Use base ifelse instead of stricter dplyr::if_else
While it is good to have a strict check of character types, when the
HS>FCL map is modified with the addition of new codes (previously
unmapped) right now there is no check that the mapping has the expected
format. Thus, while this solution works, it would be in any case have
a set of tests that check whether all inputs have the required
classes/formats. This has to do with #132.